se cambio el tipo de dato de la latitud y longitud , de f32 a f64

+++ src/apps/generar_camaras.rs
@@ -16,8 +16,8 @@ fn main() -> io::Result<()> {
     for i in 0..10 {
         let id: u8 = i;
         let mut rng = rand::thread_rng();
-        let latitude: f32 = rng.gen_range(-34.6047344..-34.6027350);
-        let longitude: f32 = rng.gen_range(-58.3871838..-58.3851850);
+        let latitude: f64 = rng.gen_range(-34.6047344..-34.6027350);
+        let longitude: f64 = rng.gen_range(-58.3871838..-58.3851850);
 
         let range: u8 = 3;
         let border_cameras: u8 = rng.gen_range(0..9); /* Ver */
